+++ CHANGES.md
@@ -1,5 +1,6 @@
 # 1.1.0
 * Get Categories (V1) is now fully depricated
+* Added new Activities API - Assign to specific engagement the list of activities
 
 ## 1.0.1
 * added a possibility of usage a custom http client